Improvements
Added resistor from pin 1 of regulator to ground to allow for Adj regulator to be used. Defaulted to 0 ohm.

Added 0 ohm resistor to voltage output. This allows a resistance to be added in to protect SAOs with no current limiting resistor on LEDs.

Tied USB shield to ground using 1M resistor and 0.1uF cap.